+++ CHANGELOG.md
@@ -121,6 +121,7 @@ Earthquake source parameters from P- or S-wave displacement spectra
   than 500 km)
 - Fix a bug in generating evid form origin time when reading origin time from
   SAC header and the number of seconds was 59
+- Fix a crash when no map tiles were available at the selected zoom level
 - Fix `source_residuals`, which didn't work anymore
 
 ## v1.6 - 2022-08-02